requirements.txt
python-telegram-bot
openai
elevenlabs
python-dotenv
